+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# [v0.19.1](https://github.com/upb-uc4/hyperledger_chaincode/compare/v0.19.0...v0.19.1) (2021-03-02)
+
+## Feature
+- check transactions for timestamp validity
+
+
+## Refactor
+- consolidate transactions and outsource transaction validation
+
 # [v0.19.0](https://github.com/upb-uc4/hyperledger_chaincode/compare/v0.18.2...v0.19.0) (2021-03-01)
 
 ## Refactor